tao.Keys manages multiple certs; cleanup tls
tao.Keys can now keep track of multiple x509 certs.
Move common tls setup code into a single place.
Add new MessageStream features (used in TaoCA).

+++ go/apps/tao_admin/tao_admin.go
@@ -674,7 +674,7 @@ func createUserKeys() {
 
 	subjectName := tao.NewX509Name(&xd)
 	userKeyPath := *options.String["user_key_path"]
-	_, err = tao.NewSignedOnDiskPBEKeys(tao.Signing, upwd, userKeyPath, subjectName, int(xd.GetSerialNumber()), policyKey)
+	_, err = tao.NewSignedOnDiskPBEKeys(tao.Signing, upwd, userKeyPath, subjectName, int64(xd.GetSerialNumber()), policyKey)
 	options.FailIf(err, "Can't create user signing key")
 }
